+++ src/be/ac/ulb/infof307/g05/ProjectStruct.java
@@ -2,6 +2,6 @@ package be.ac.ulb.infof307.g05;
 
 public class ProjectStruct {
 	public String name;
-	public float width;
-	public float length;
+	public int width;
+	public int length;
 }
